Print the way number is pronounced

+++ piscine42/rush02/include/rush02.h
@@ -6,7 +6,7 @@
 /*   By: vietnguy <vietnguy@student.42singapore.sg  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/07 23:13:42 by vietnguy          #+#    #+#             */
-/*   Updated: 2023/07/08 17:11:23 by vietnguy         ###   ########.fr       */
+/*   Updated: 2023/07/08 17:54:38 by vietnguy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 
 typedef struct s_record
 {
-	int	key;
-	char	*val;
+	long long	key;
+	char		*val;
 }	t_record;
 /* parse_dict.c */
 t_record	*ft_parse_dict(char *filepath);
@@ -34,7 +34,7 @@ void	ft_strcpylr(char *dest, char *src, int l, int r);
 void	ft_strcpy(char *dest, char *src);
 /* strconv.c */
 int		ft_valid_number(char *str);
-int		ft_atoi(char *str);
+long long	ft_atoi(char *str);
 /* strsplit.c */
 int	issep(char c, char *sep);
 int	ft_count_words(char *str, char *sep);
@@ -46,8 +46,8 @@ int	ft_dictkeycmp(t_record k1, t_record k2);
 void	ft_sortdict(t_record *dict);
 /* dict.c */
 int	ft_dictlen(t_record *dict);
-char	*ft_get_val(t_record *dict, int key);
-void	ft_eval_key(t_record *dict, int key);
+char	*ft_get_val(t_record *dict, long long key);
+void	ft_eval_key(t_record *dict, long long key);
 void	ft_free_dict(t_record *dict);
 void	ft_print_dict(t_record *dict);
 #endif
